Rename send to send_payload
Defining the `send` method overrides the `send` on Ruby's Object
class (https://ruby-doc.org/core-2.4.5/Object.html#method-i-send).
This normally wouldn't be a problem, but Fluentd uses `Object.send`
in their `RootAgent` to call the lifecycle functions, which causes
a stacktrace dump on shutdown by trying to call `send` with a symbol.